requirements.txt
django
django-npm
pyyaml
web3
channels-redis